requirements.txt
requests~=2.28.1
responses~=0.21.0
python-dateutil
tenacity
requests-mock
vcrpy
freezegun